added combined motor power set block
this allows atomically setting power to both motors. This is needed to allow
the robot predictably move directly forward or backward on slow execution speeds.
Without this the robot would turn on one side before the other which would cause
the robot to list unhelpfully to one side or the other.

Fixes #144