Show minimal blockly code at the first launch
Create a new, minimalistic, blockly program that just sets the motor
power on one side.

Make this new program the default code that gets loaded on the very
first start of the simulator.

Note that the current logic of handling blockly code (persisting code
in local storage across re-loads and challenges etc) remains the same.

Resolves #148