Add limit to number of blocks
Fix #92 by allow challange config to specify the number of blocks allowed
to be used and whether the limit is soft (not enforced) or hard (enforced).
This should help students to identify the are on the wrong
path,

There is a feedback of how many blocks are allowed and if you within
the limit or not.