Show motors sensors from the robot
Before this change the information about the motors in RobotView
was captured when setting the motors and not from the robot itself.

This now uses the RobotSpec and SimRobot, like the rest of the sensors,
to decide what to present and get the actual values of the motors.

We force RobotView to update internally because now we do not write
to redux on every change to the motor values which used to force all
components to refresh before.

NOTE: I've changed the version of `@frud/simulator-core` to be pinned to make sure
it will not get updated by mistake by Dependabot.